edit models of villa

+++ RestBox/Villas/models.py
@@ -8,6 +8,6 @@ class Villa(models.Model):
     title = models.CharField(max_length=200, verbose_name="title")
     address = models.TextField(verbose_name="address")
     capacity = models.IntegerField(verbose_name="capacity")
-    price_per_night = models.IntegerField(max_digits=10, verbose_name="price_per_night")
+    price_per_night = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="price_per_night")    
     def __str__(self):
         return self.title
